requirements.txt
django
drf-spectacular
djangorestframework
psycopg
django-cors-headers
whitenoise
gunicorn
apscheduler